backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
python-multipart==0.0.6
email-validator==2.1.0
pydantic-settings==2.0.3
psycopg2-binary==2.9.9

# Security enhancements
slowapi==0.1.9
limits==5.4.0

# Logging enhancements
structlog==25.4.0
python-json-logger==2.0.7

# Additional security
cryptography>=41.0.0
bcrypt>=4.0.0

# Development tools (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

